# llm-verifier/api/docs/swagger.yaml
definitions:
  analytics.TaskRequirements:
    properties:
      budget_limit:
        description: Maximum cost per request
        type: number
      complexity:
        description: simple, medium, complex
        type: string
      context_length:
        description: Required context window
        type: integer
      required_features:
        description: json_mode, function_calling, etc.
        items:
          type: string
        type: array
      speed_requirement:
        description: fast, normal, quality_first
        type: string
      task_type:
        description: coding, writing, analysis, etc.
        type: string
    type: object
...
  api.LoginRequest:
    properties:
      password:
        maxLength: 100
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /analytics/compare:
    post:
      consumes:
      - application/json
      description: Compares multiple models based on task requirements
      parameters:
      - description: Comparison request
        in: body
        name: comparison
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Model comparison
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Compare models
      tags:
      - analytics
  /analytics/cost:
    get:
      description: Retrieves cost optimization analysis and recommendations
      parameters:
      - description: 'Time range in days (default: 30)'
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cost analysis data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Get cost analysis
      tags:
      - analytics
  /analytics/insights:
    get:
      description: Retrieves usage insights and optimization recommendations
      parameters:
      - description: 'Time range in days (default: 30)'
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usage insights
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Get usage insights
      tags:
      - analytics
  /analytics/recommendations:
    post:
      consumes:
      - application/json
      description: Retrieves AI-powered model recommendations based on task requirements
      parameters:
      - description: Task requirements
        in: body
        name: requirements
        required: true
        schema:
          $ref: '#/definitions/analytics.TaskRequirements'
      produces:
      - application/json
      responses:
        "200":
          description: Model recommendations
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Get model recommendations
      tags:
      - analytics
  /analytics/trends:
    get:
      description: Retrieves performance trend analysis with anomaly detection and
        forecasting
      parameters:
      - description: Metric name (e.g., response_time, throughput)
        in: query
        name: metric
        required: true
        type: string
      - description: 'Time range in hours (default: 24)'
        in: query
        name: hours
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Trend analysis data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Get performance trends
      tags:
      - analytics
  /analytics/usage:
    get:
      description: Retrieves usage pattern analysis including hourly/daily patterns
        and recommendations
      parameters:
      - description: 'Time range in days (default: 7)'
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usage analysis data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Get usage patterns
      tags:
      - analytics
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login response with JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication failed
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - auth
  /config/export:
    post:
      consumes:
      - application/json
      description: Exports system configuration in various formats for AI CLI tools
      parameters:
      - description: Export configuration
        in: body
        name: export
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Exported configuration
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Export configuration
      tags:
      - config
...
  /issues:
    get:
      description: Retrieves system issues and problems with optional filtering
      parameters:
      - description: Number of results per page
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by issue type
        in: query
        name: issue_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of issues
          schema:
            additionalProperties: true
            type: object
      summary: Get issues
      tags:
      - issues
  /models:
    get:
      description: Retrieves a list of all LLM models with pagination
      parameters:
      - description: 'Number of results per page (default: 50)'
        in: query
        name: limit
        type: integer
      - description: 'Offset for pagination (default: 0)'
        in: query
        name: offset
        type: integer
      - description: Filter by provider name
        in: query
        name: provider
        type: string
      - description: Filter by model status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of models
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameters
          schema:
            additionalProperties: true
            type: object
      summary: Get models
      tags:
      - models
    post:
      consumes:
      - application/json
      description: Creates a new LLM model (admin only)
      parameters:
      - description: Model data
        in: body
        name: model
        required: true
        schema:
          $ref: '#/definitions/api.CreateModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created model
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden (admin only)
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create model
      tags:
      - models
...
  /notifications/subscribers:
    post:
      consumes:
      - application/json
      description: Creates a new notification subscriber for specified event types
        and channels
      parameters:
      - description: Subscriber configuration
        in: body
        name: subscriber
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Subscriber created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Create notification subscriber
      tags:
      - notifications
  /notifications/test:
    post:
      consumes:
      - application/json
      description: Sends a test notification to specified channels for testing purposes
      parameters:
      - description: Test notification configuration
        in: body
        name: test
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Test notification sent
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
      summary: Send test notification
      tags:
      - notifications
...
  /system/database-stats:
    get:
      description: Returns database performance and size statistics
      produces:
      - application/json
      responses:
        "200":
          description: Database statistics
          schema:
            additionalProperties: true
            type: object
        "503":
          description: Database not available
          schema:
            additionalProperties: true
            type: object
      summary: Get database statistics
      tags:
      - system
...
swagger: "2.0"